fix: type 재정리 & submit 함수 이동
