+++ apps/web/src/components/auction-detail/bid-form/bottom.tsx
@@ -1,54 +1,14 @@
 'use client';
 
-import { useState } from 'react';
-
 import { ButtonFavorite } from '@/components/auction-detail/';
 import { Button } from '@/components/ui';
 
 import { formatCurrencyWithUnit } from '@/lib/utils/price';
 
-import { BidFormBottomProps } from '@/types/bid';
+import { BidBaseProps } from '@/types/bid';
 
-const BidFormBottom = ({
-  auctionId,
-  currentPrice,
-  isExpired,
-  isBidding,
-  onChange,
-  bidTableRef,
-}: BidFormBottomProps) => {
-  const [bidPrice, setBidPrice] = useState<number | null>(null);
-
+const BidFormBottom = ({ auctionId, currentPrice, isExpired }: BidBaseProps) => {
   const formattedCurrentPrice = formatCurrencyWithUnit(currentPrice);
-
-  const handleBidClick = (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!isBidding) {
-      onChange(true);
-      return;
-    }
-
-    // 경매가 종료된 경우 입찰 불가
-    if (isExpired) {
-      alert('경매가 종료되어 입찰할 수 없습니다.');
-      return;
-    }
-
-    if (bidPrice === null) return alert('입찰 금액을 입력해주세요');
-
-    // 여기에 실제 입찰 로직 추가
-    try {
-      const results = true;
-
-      //입찰 완료 시 bid-table로 scroll 이동
-      if (results) {
-        bidTableRef.current?.scrollIntoView({ behavior: 'smooth', block: 'start' });
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
 
   return (
     <div className="z-40 flex items-center gap-2 bg-white px-5 pb-6 pt-3">
@@ -62,7 +22,6 @@ const BidFormBottom = ({
       <Button
         size="medium"
         color={isExpired ? 'disabled' : 'primary'}
-        onClick={handleBidClick}
         disabled={isExpired}
         className={`w-1/3`}
       >
